requirements.txt
streamlit
psycopg2-binary
python-dotenv
pandas
plotly
jinja2
openai
langchain-core
langchain-groq
langchain-community
SQLAlchemy
tabulate
bson
pymango
urllib.parse
pymongo[srv]
